server/requirements.txt
fastapi
uvicorn
redis
kafka-python